refactor nummerical cell types
- change namespaces into structs
- update traits spezilization of `FieldPosition`
